+++ sigil/domain/campaign/actions.py
@@ -1,12 +1,12 @@
 """Use cases for campaign module"""
 
-from sigil.store.interfaces import CampaignStore
+from sigil.storage.interfaces import CampaignStorage
 
 
 class CampaignManager:
-    def __init__(self, store: CampaignStore):
-        self.store = store
+    def __init__(self, storage: CampaignStorage):
+        self.storage = storage
 
     async def list_all(self):
-        campaigns = await self.store.list()
+        campaigns = await self.storage.list()
         return campaigns