+++ sigil/storage/actions.py
@@ -1,0 +1,3 @@
+from sigil.storage.adapters.psql.commands import init_storage as init
+from sigil.storage.adapters.psql.commands import upgrade_storage as upgrade
+from sigil.storage.seed import seed