+++ sigil/storage/adapters/psql/alembic/env.py
@@ -5,8 +5,8 @@ from alembic import context
 from sqlalchemy import engine_from_config, pool
 from sqlalchemy.ext.asyncio import AsyncEngine
 
-from sigil.store.adapters.psql import get_connection_string
-from sigil.store.adapters.psql.models import Base
+from sigil.storage.adapters.psql import get_connection_string
+from sigil.storage.adapters.psql.models import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.