+++ sigil/storage/adapters/psql/commands.py
@@ -4,9 +4,9 @@ from os import path
 from alembic import command
 from alembic.config import Config
 
-from sigil.store.adapters.psql import async_session, engine
-from sigil.store.adapters.psql.models.base import Base
-from sigil.store.interfaces import CampaignStore
+from sigil.storage.adapters.psql import async_session, engine
+from sigil.storage.adapters.psql.models.base import Base
+from sigil.storage.interfaces import CampaignStorage
 
 dirname = path.dirname(__file__)
 alembic_ini = path.join(dirname, "alembic.ini")
@@ -44,10 +44,10 @@ def include_storages(func):
     async def _include_storages(*args, **kwargs):
         async with async_session() as session:
             async with session.begin():
-                stores = {
-                    CampaignStore.__name__: CampaignStore(session),
+                storages = {
+                    CampaignStorage.__name__: CampaignStorage(session),
                 }
-                await func(*args, **kwargs, stores=stores)
+                await func(*args, **kwargs, storages=storages)
 
                 await session.commit()
         await engine.dispose()