+++ sigil/storage/adapters/psql/models/__init__.py
@@ -1,0 +1,3 @@
+from sigil.storage.adapters.psql.models.base import Base
+
+from .campaign import CampaignModel