+++ sigil/storage/adapters/psql/storages/campaign.py
@@ -4,12 +4,12 @@ from pydantic import UUID4
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from sigil.domain.campaign.entities import Campaign
-from sigil.store.adapters.psql import select
-from sigil.store.adapters.psql.models import CampaignModel
-from sigil.store.domain.campaign import BaseCampaignStore
+from sigil.storage.adapters.psql import select
+from sigil.storage.adapters.psql.models import CampaignModel
+from sigil.storage.domain.campaign import BaseCampaignStorage
 
 
-class CampaingStorePsql(BaseCampaignStore):
+class CampaingStoragePsql(BaseCampaignStorage):
     def __init__(self, session: AsyncSession):
         self.session = session