+++ sigil/store/adapters/psql/models/__init__.py
@@ -1,3 +1,0 @@
-from sigil.store.adapters.psql.models.base import Base
-
-from .campaign import CampaignModel